refactor: 카드오픈과 관련된 클래스 위치이동
