+++ src/test/java/blackjack/domain/player/DealerTest.java
@@ -5,6 +5,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardNumber;
 import blackjack.domain.card.CardShape;
+import blackjack.domain.player.cardopen.AllCardsOpenStrategy;
+import blackjack.domain.player.cardopen.OneCardOpenStrategy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
